Fixes an issue in the VexFlow library that has to do with drawing doubled notes in the same StaveNote (i.e. F and F# or G and G#).
The issue is that only one note is being rendered instead of two, or one
note is being rendered on top of the other, giving the appearance of
only one note. What should happen is that the second note on the same
line should be displaced so that it is visible. This change to the
StaveNote draw() method does just that.

IMPORTANT NOTE: this is a local modification to the the VexFlow library
that has *not* been contributed back upstream at this time, although the
goal is to get the issue fixed so that this workaround is not necessary.