Update harmony.css
Improvements to the appearance of the header and notifications area. Two small edits to the html structure and content should accompany this change in the css:
(1) The element with id="notificationAlerts" should be moved up so that it precedes the element containing the span which reads "Help."
(2) The button reading "View" in notifications should be changed to "Expand or collapse".