Update midi.js
Behavior of sustain pedals is apparently inconsistent across devices. The new variable `interpret_low_as_on` needs to be reworked as an option set in the UI. Moreover, the broadcasting of pedal events needs to be limited to instances where a change of state is occurring. Yahama DGX-620 for example will bank two chords every time sustain pedal is activated.